Add mutating webhook for app CRs
