+++ main.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/giantswarm/app-admission-controller/config"
 	"github.com/giantswarm/app-admission-controller/pkg/app"
-	"github.com/giantswarm/app-admission-controller/pkg/aws/g8scontrolplane"
 	"github.com/giantswarm/app-admission-controller/pkg/mutator"
 	"github.com/giantswarm/app-admission-controller/pkg/validator"
 )
@@ -23,7 +22,7 @@ func main() {
 		panic(microerror.JSON(err))
 	}
 
-	g8scontrolplaneMutator, err := g8scontrolplane.NewMutator(config)
+	appMutator, err := app.NewMutator(config)
 	if err != nil {
 		panic(microerror.JSON(err))
 	}
@@ -35,7 +34,7 @@ func main() {
 
 	// Here we register our endpoints.
 	handler := http.NewServeMux()
-	handler.Handle("/mutate/g8scontrolplane", mutator.Handler(g8scontrolplaneMutator))
+	handler.Handle("/mutate/app", mutator.Handler(appMutator))
 	handler.Handle("/validate/app", validator.Handler(appValidator))
 
 	handler.HandleFunc("/healthz", healthCheck)
