+++ pkg/unittest/default_k8sclient.go
@@ -2,8 +2,6 @@ package unittest
 
 import (
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v2/pkg/apis/infrastructure/v1alpha2"
-	releasev1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/release/v1alpha1"
 	"github.com/giantswarm/apiextensions/v2/pkg/clientset/versioned"
 	fakeg8s "github.com/giantswarm/apiextensions/v2/pkg/clientset/versioned/fake"
 	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
@@ -14,7 +12,6 @@ import (
 	"k8s.io/client-go/kubernetes"
 	fakek8s "k8s.io/client-go/kubernetes/fake"
 	"k8s.io/client-go/rest"
-	"sigs.k8s.io/cluster-api/api/v1alpha2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
@@ -36,18 +33,6 @@ func FakeK8sClient() k8sclient.Interface {
 			panic(err)
 		}
 
-		err = v1alpha2.AddToScheme(scheme)
-		if err != nil {
-			panic(err)
-		}
-		err = infrastructurev1alpha2.AddToScheme(scheme)
-		if err != nil {
-			panic(err)
-		}
-		err = releasev1alpha1.AddToScheme(scheme)
-		if err != nil {
-			panic(err)
-		}
 		_ = fakek8s.AddToScheme(scheme)
 		client := fakek8s.NewSimpleClientset()
 		g8sclient := fakeg8s.NewSimpleClientset()
